Add utils.ts
