+++ packages/akashic-cli-export/spec/src/zip/convertSpec.ts
@@ -7,7 +7,6 @@ import {
 	bundleScripts,
 	convertGame,
 	ConvertGameParameterObject,
-	checkAudioAssetExtensions,
 	validateGameJsonForNicolive
 } from "../../../lib/zip/convert";
 
@@ -430,10 +429,12 @@ describe("convert", () => {
 					expect(gameJson.globalScripts.includes("node_modules/@hoge/testmodule/lib/ModuleC.js")).toBeFalsy();
 					expect(gameJson.globalScripts.includes("node_modules/@hoge/testmodule/lib/index.js")).toBeFalsy();
 					// バンドルに含まれず削除された利用されていない不要なファイルをログ出力
+					/* eslint-disable max-len */
 					expect(consoleSpy).toHaveBeenCalledWith("excluded node_modules/@hoge/testmodule/lib/ModuleB.js due to unreachable/unhandled.");
 					expect(consoleSpy).toHaveBeenCalledWith("excluded node_modules/@hoge/testmodule/lib/ModuleC.js due to unreachable/unhandled.");
 					expect(consoleSpy).toHaveBeenCalledWith("excluded node_modules/@hoge/testmodule/lib/index.js due to unreachable/unhandled.");
 					expect(consoleSpy).toHaveBeenCalledWith("excluded script/bar.js due to unreachable/unhandled.");
+					/* eslint-enable max-len */
 					done();
 				}, done.fail);
 		});
@@ -634,68 +635,5 @@ describe("game.json validation with nicolive option", () => {
 	it("nicolive property does not exist", () => {
 		delete gamejson.environment.nicolive;
 		expect(() => validateGameJsonForNicolive(gamejson)).toThrow();
-	});
-});
-
-describe("checkAudioAssetExtensions", () => {
-	const gamejson: any = {
-		assets: {
-			audio1: {
-				type: "audio",
-				duration: 456,
-				systemId: "sound",
-				global: true,
-				path: "assets/audio/se/audio1",
-				hint: {
-				  extensions: [".ogg", ".m4a"]
-				}
-			},
-			audio2: {
-				type: "audio",
-				duration: 456,
-				systemId: "sound",
-				global: true,
-				path: "assets/audio/se/audio2",
-				hint: {
-				  extensions: [".m4a"]
-				}
-			},
-			audio3: {
-				type: "audio",
-				duration: 456,
-				systemId: "sound",
-				global: true,
-				path: "assets/audio/se/audio3",
-				hint: {
-				  extensions: [".ogg"]
-				}
-			},
-			audio4: {
-				type: "audio",
-				duration: 456,
-				systemId: "sound",
-				global: true,
-				path: "assets/audio/se/audio4",
-				hint: {
-				  extensions: [".ogg", ".aac"]
-				}
-			},
-			nohint: {
-				type: "audio",
-				duration: 456,
-				systemId: "sound",
-				path: "assets/audio/se/nohint",
-				global: true
-			}
-		}
-	};
-
-	it("Log output when checkAudioAssetExtensions", () => {
-		const spy = jest.spyOn(global.console, "warn");
-		checkAudioAssetExtensions(gamejson);
-
-		expect(spy).toHaveBeenCalledWith("may be no sound depending on the environment because no .ogg file in assets/audio/se/audio2.");
-		expect(spy).toHaveBeenCalledWith("may be no sound depending on the environment because no .m4a or .aac file in assets/audio/se/audio3.");
-		spy.mockRestore();
 	});
 });
